Extract the protocol-agnostic proxy core into proxy.core
The single-stable-port listener, session registry, back-pressure, expiry
and shutdown had all lived in PostgresProxyServer, tangled with the
Postgres wire protocol. Pull that lifecycle machinery into a new
proxy.core package as a generic ProxyServer driven by a ProxyProtocol
strategy, so a second protocol (MySQL/MariaDB) can plug into the same
listener without duplicating it.

core now owns ProxyServer, ProxySession (protocol-agnostic: it carries
the raw upstream coordinates, not a Postgres target factory),
ProxyProtocol + AuthenticatedClient, the ProxyConnection relay
interface, the TLS helpers (incl. enableSSL), writeAndFlush,
INFINITE_ACCESS and getShutdownDate. The Postgres half becomes
PostgresProtocol: its handshake() wraps the existing SASL/startup
authentication and its connect() opens the upstream and builds the
Connection relay (which now implements ProxyConnection).

Pure refactor: no behavior change. The proxy suites are unchanged apart
from constructing ProxyServer(port, PostgresProtocol(...)) instead of
PostgresProxyServer(...).
