+++ backend/src/main/kotlin/dev/kviklet/kviklet/proxy/core/OutputStreamExtension.kt
@@ -1,5 +1,5 @@
 // This file is not MIT licensed
-package dev.kviklet.kviklet.proxy.postgres
+package dev.kviklet.kviklet.proxy.core
 
 import java.io.OutputStream
