+++ backend/src/main/kotlin/dev/kviklet/kviklet/proxy/postgres/ProxyServerConfig.kt
@@ -1,6 +1,9 @@
 // This file is not MIT licensed
 package dev.kviklet.kviklet.proxy.postgres
 
+import dev.kviklet.kviklet.proxy.core.ProxyServer
+import dev.kviklet.kviklet.proxy.core.TlsCertEnvConfig
+import dev.kviklet.kviklet.proxy.core.tlsCertificateFactory
 import dev.kviklet.kviklet.service.EventService
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.context.annotation.Bean
@@ -12,7 +15,7 @@ class ProxyServerConfig {
     // A single long-lived Postgres proxy listener, bound at application startup on the configured stable
     // port (default 5432). destroyMethod closes it on context shutdown. A port <= 0 disables the listener
     // (start() is a no-op); the test profile sets -1 so the bean never binds a real port during the Spring
-    // integration tests, which spin up their own PostgresProxyServer instances on random ports instead.
+    // integration tests, which spin up their own ProxyServer instances on random ports instead.
     // (The e2e stack does bind the default port, harmlessly, inside its own container network.)
     //
     // @Lazy(false) forces eager creation so the port really binds at boot: the app sets
@@ -23,8 +26,8 @@ class ProxyServerConfig {
         eventService: EventService,
         tlsCertConfig: TlsCertEnvConfig,
         @Value("\${kviklet.proxy.postgres.port:5432}") port: Int,
-    ): PostgresProxyServer {
-        val server = PostgresProxyServer(port, eventService, tlsCertificateFactory(tlsCertConfig))
+    ): ProxyServer {
+        val server = ProxyServer(port, PostgresProtocol(eventService, tlsCertificateFactory(tlsCertConfig)))
         server.start()
         return server
     }
