+++ backend/src/main/resources/application-test.properties
@@ -33,6 +33,6 @@ spring.datasource.driver-class-name=org.testcontainers.jdbc.ContainerDatabaseDri
 server.tomcat.accesslog.enabled=true
 server.tomcat.accesslog.pattern=%t %a "%r" %s (%D ms)
 app.in-docker=false
-# Disable the boot-bound proxy listener in tests; the proxy suites start their own PostgresProxyServer
+# Disable the boot-bound proxy listener in tests; the proxy suites start their own ProxyServer
 # instances on random ports instead.
 kviklet.proxy.postgres.port=-1
