+++ backend/src/test/kotlin/dev/kviklet/kviklet/proxy/PostgresProxyConnectionLifecycleTest.kt
@@ -3,10 +3,10 @@ package dev.kviklet.kviklet.proxy
 
 import dev.kviklet.kviklet.db.EventAdapter
 import dev.kviklet.kviklet.db.ExecutionRequestAdapter
+import dev.kviklet.kviklet.proxy.core.ProxyServer
 import dev.kviklet.kviklet.proxy.helpers.ProxyInstance
 import dev.kviklet.kviklet.proxy.helpers.directConnectionFactory
 import dev.kviklet.kviklet.proxy.helpers.proxyServerFactory
-import dev.kviklet.kviklet.proxy.postgres.PostgresProxyServer
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertTrue
@@ -90,11 +90,7 @@ class PostgresProxyConnectionLifecycleTest {
         assertEquals(expected, last, "Upstream connection count on the target database did not settle")
     }
 
-    private fun awaitConnectionCount(
-        expected: Int,
-        target: PostgresProxyServer = proxy.proxy,
-        timeoutMillis: Long = 20_000,
-    ) {
+    private fun awaitConnectionCount(expected: Int, target: ProxyServer = proxy.proxy, timeoutMillis: Long = 20_000) {
         val deadline = System.currentTimeMillis() + timeoutMillis
         while (System.currentTimeMillis() < deadline && target.currentConnections != expected) {
             Thread.sleep(100)
@@ -104,7 +100,7 @@ class PostgresProxyConnectionLifecycleTest {
 
     private fun awaitPendingHandshakeCount(
         expected: Int,
-        target: PostgresProxyServer = proxy.proxy,
+        target: ProxyServer = proxy.proxy,
         timeoutMillis: Long = 20_000,
     ) {
         val deadline = System.currentTimeMillis() + timeoutMillis
@@ -119,7 +115,7 @@ class PostgresProxyConnectionLifecycleTest {
     // verified here by reflection rather than exposing a production counter that would sit confusingly
     // next to currentConnections.
     private fun trackedConnectionCount(): Int {
-        val field = PostgresProxyServer::class.java.getDeclaredField("clientConnections")
+        val field = ProxyServer::class.java.getDeclaredField("clientConnections")
         field.isAccessible = true
         return (field.get(proxy.proxy) as Collection<*>).size
     }
